moved toc into separate component
